project3_dashboard/requirements.txt
flask==3.0.0
flask-cors==4.0.0
psycopg2-binary==2.9.9
pandas==2.1.3
scikit-learn==1.3.2
beautifulsoup4==4.12.2
